refactor: extract IssueExportReviewItem.swift from IssueExportReview.swift
Splits the `IssueExportReviewItem` struct (per-issue review record with
matches + resolution) out of IssueExportReview.swift so that file holds
only the top-level IssueExportReview aggregate.

Byte-identical move of lines 53-108. Both files import Foundation.

IssueExportReview.swift: 130 → 73 lines. Tests: 667.
